db/requirements.txt
python-dotenv
sqlalchemy
psycopg2-binary
pgvector
alembic
